// recon/src/app/shared/services/enum.ts
export enum ReconDataType {
  RECONCILED = 'RECONCILED',
  UNRECONCILED = 'UNRECONCILED',
  EDIT_INVALID_DATA = 'EDIT_INVALID_DATA',
  MASTER_DATA = 'MASTER_DATA',
  KNOCK_OFF = 'KNOCK_OFF',
  RELAX_MATCH = 'RELAX_MATCH',
  PROPOSE_MATCH = 'PROPOSE_MATCH',
  DUPLICATE_MATCH_KNOCK_OFF = 'DUPLICATE_MATCH_KNOCK_OFF'
}